# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ECM Platform Copilot API
  description: >
    REST surface for East Coast Mechanical's CRM ("ECM Platform").
    This spec has two parts:
      1. **Existing endpoints** (`/api/stripe/webhook`, `/api/sms/webhook`, `/api/cron/email-reminders`) —
         documented as they exist today. These are single-purpose webhooks/cron jobs and are
         NOT suitable for Copilot Studio to call.
      2. **Copilot Studio endpoint** (`/api/copilot/agent`) — the integration point for
         Microsoft Copilot Studio, wrapping the internal `respondToInternalChat` agent
         (lib/internal-assistant/agent.ts) behind bearer-token auth. Requires two headers:
         `Authorization` (the COPILOT_API_KEY, as literal value "Bearer <key>") and
         `x-vercel-protection-bypass` (to get past this project's Vercel SSO wall).
  version: "0.1.0"
servers:
  - url: https://ecm-crm-eastcoastmechanics-projects.vercel.app
    description: >
      Production (Vercel). Use this alias, not the bare ecm-crm.vercel.app
      domain -- that one is not kept in sync with production deploys and
      serves stale/404 responses.
paths:
  /api/copilot/agent:
    post:
      summary: Natural-language CRM command endpoint for Copilot Studio
      description: >
        Single freeform entry point for Copilot Studio. Accepts one line of natural language,
        classifies it via Claude into create_customer / create_document / create_task / unclear,
        executes the matching action against Supabase, and returns a plain-text reply.
        This is a thin HTTP wrapper around the existing `runAssistantCommand` function —
        no new business logic, only a new transport + auth layer.
      operationId: runAssistantCommand
      security:
        - bearerAuth: []
          vercelProtectionBypass: []
      parameters:
        - name: x-vercel-protection-bypass
          in: header
          required: true
          description: >
            Required in addition to the bearer token. This project has Vercel
            Authentication (SSO) enabled on all *.vercel.app URLs, which
            blocks anonymous callers before the request reaches app code.
            This header's value must match the VERCEL_AUTOMATION_BYPASS_SECRET
            configured under Project Settings -> Deployment Protection ->
            Protection Bypass for Automation.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command:
                  type: string
                  description: One-line natural-language instruction, e.g. "Add a customer named Jane Doe, phone 555-1234"
                  example: "Create an estimate for John Smith - replacing a leaking condenser at 12 Main St"
      responses:
        "200":
          description: Command processed (successfully or with a clarifying/error reply — check `reply` text)
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
                    example: "Added customer Jane Doe."
        "401":
          description: Missing or invalid bearer token
        "400":
          description: Missing `command` in request body
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        For /api/copilot/agent and /api/cron/email-reminders: a dedicated secret
        (COPILOT_API_KEY or CRON_SECRET). Because this is modeled as an apiKey header
        (Swagger 2.0 has no native bearer-scheme type), the value entered when configuring
        this connector must be the literal string "Bearer <secret>", including the "Bearer " prefix.
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
    hmacSignature:
      type: apiKey
      in: header
      name: x-signature
    vercelProtectionBypass:
      type: apiKey
      in: header
      name: x-vercel-protection-bypass